+++ crates/ruff_linter/resources/test/fixtures/pylint/literal_membership.py
