+++ crates/ruff_linter/src/rules/pylint/rules/mod.rs
@@ -24,6 +24,7 @@ pub(crate) use invalid_envvar_value::*;
 pub(crate) use invalid_str_return::*;
 pub(crate) use invalid_string_characters::*;
 pub(crate) use iteration_over_set::*;
+pub(crate) use literal_membership::*;
 pub(crate) use load_before_global_declaration::*;
 pub(crate) use logging::*;
 pub(crate) use magic_value_comparison::*;
@@ -39,7 +40,6 @@ pub(crate) use repeated_equality_comparison::*;
 pub(crate) use repeated_isinstance_calls::*;
 pub(crate) use return_in_init::*;
 pub(crate) use self_assigning_variable::*;
-pub(crate) use set_membership::*;
 pub(crate) use single_string_slots::*;
 pub(crate) use subprocess_popen_preexec_fn::*;
 pub(crate) use subprocess_run_without_check::*;
@@ -87,6 +87,7 @@ mod invalid_envvar_value;
 mod invalid_str_return;
 mod invalid_string_characters;
 mod iteration_over_set;
+mod literal_membership;
 mod load_before_global_declaration;
 mod logging;
 mod magic_value_comparison;
@@ -102,7 +103,6 @@ mod repeated_equality_comparison;
 mod repeated_isinstance_calls;
 mod return_in_init;
 mod self_assigning_variable;
-mod set_membership;
 mod single_string_slots;
 mod subprocess_popen_preexec_fn;
 mod subprocess_run_without_check;
